moved passport.js into a new authentication directory and updated server.js to link to it
